backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
pydantic
passlib[bcrypt]
bcrypt==3.2.0           # <--- CRITICAL: Keeps the compatible version
python-multipart
python-jose[cryptography]
email-validator